Update Menu with Buttons and Colliders

+++ Virus_Royale/Virus_Royale/colisionManager.cpp
@@ -5,12 +5,12 @@ namespace Collisions
 	void CheckBounds(Vector2& pos) 
 	{
 		if (pos.x < 0)
-			pos.x = GetScreenWidth();
-		if (pos.x > GetScreenWidth())
+			pos.x = (float)GetScreenWidth();
+		if (pos.x > (float)GetScreenWidth())
 			pos.x = 0;
 		if (pos.y < 0)
-			pos.y = GetScreenHeight();
-		if (pos.y > GetScreenHeight())
+			pos.y = (float)GetScreenHeight();
+		if (pos.y > (float)GetScreenHeight())
 			pos.y = 0;
 	}
